Renaming rich_converter to card_rich and adding a test script for it.
